Fix index assignation at enqueueing
Index was previously set before enqueue for the first node. That could
cause a data race because two concurrent threads could have the same
index, the second one being enqueued after an enqueue+dequeue of the
first one.

Now index is set after enqueueing the node, and the queue index is
used when the prev index is not available, with a double check to
ensure synchronization.